requirements.txt
ipython-autotime==0.3.1
kfp==2.6.0
google-cloud-pipeline-components==2.12.0
streamlit==1.34.0
pandas==2.2.2
numpy==1.26.4
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.17.2
python-dotenv==1.0.1
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.22.0
tabulate==0.9.0
protobuf==4.25.7